Add mDNS domain only hosting
* Add homepage and enable Avahi mDNS

* Add Ansible playbooks for post install configuration

* Do not enable backup by default

Backups do not work until a user provides an rclone configuration.
This keeps the services from showing up as failed.

* Add some extra hosts for homepage

* Remove non-functional site-monitors from homepage

* Add cleaned path for jellyfin logs

* Add default media user and subdirectories

* Reassign UIDs and GIDs

I have to hardcode the UID and GIDs to provide to the quadlets,
but the hardcoded IDs that I selected were clashing with dynamically assigned UIDs.
This reassigns them to start with IDs at 500, which seems to be far away from the 900+ range that systemd uses.

* Add backup restore and fix up Ansible

* Wait for network before mounting NFS